// digital-wave-games/src/app/shared/models/dto/signUp/signupformDTO.ts
export interface SignUpFormDTO {
  email: string;
  username: string;
  name: string
  password: string;
  postalCode: string;
  city: string;
  state: string;
  district: string;
  number: string;
  additionalInfo: string;
  street: string;
  phone1: string;
  phone2: string;
  phone3: string;
  secondaryEmail: string;
  code: string;
}